Add test api-promise/client-class-member
