+++ src/transforms/v2-to-v3/__fixtures__/api-promise/client-class-member.output.js
@@ -1,0 +1,16 @@
+import { DynamoDB } from "@aws-sdk/client-dynamodb";
+
+// Client as class member
+class ClientClassMember {
+  constructor(clientInCtr = new DynamoDB()) {
+    this.clientInClass = clientInCtr;
+  }
+
+  async listTables() {
+    return await this.clientInClass.listTables();
+  }
+
+  async listTagsOfResource() {
+    return await this.clientInClass.listTagsOfResource({ ResourceArn: "STRING_VALUE" });
+  }
+}
