+++ src/transforms/v2-to-v3/__fixtures__/api-promise/service-import.input.js
@@ -5,21 +5,6 @@ const client = new DynamoDBClient();
 // ExpressionStatement
 client.listTables().promise();
 
-// Client as class member
-class ClientClassMember {
-  constructor(clientInCtr = new DynamoDBClient()) {
-    this.clientInClass = clientInCtr;
-  }
-
-  async listTables() {
-    return await this.clientInClass.listTables().promise();
-  }
-
-  async listTagsOfResource() {
-    return await this.clientInClass.listTagsOfResource({ ResourceArn: "STRING_VALUE" }).promise();
-  }
-}
-
 // Variable declarator
 const listTablesPromise = client.listTables().promise();
 const listTagsOfResourcePromise = client.listTagsOfResource({ ResourceArn: "STRING_VALUE" }).promise();
